backend/requirements.txt
asgiref==3.7.2
black==22.1.0
blue==0.9.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==3.2.3
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-filter==23.5
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
flake8==4.0.0
flake8-isort==6.1.1
gunicorn==21.2.0
idna==3.6
importlib_metadata==7.0.2
isort==5.13.2
mccabe==0.6.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
Petrovich==2.0.1
pillow==10.2.0
platformdirs==4.2.0
psycopg2==2.9.9
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.1.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tomli==2.0.1
tzdata==2024.1
urllib3==2.2.1
zipp==3.18.1
